Don't add hive to brains
This turned out to be a problem because we keep cloning the hive and
replacing the one returned from OpusVL::FB11::Hive. This means it's a
lot of overhead to make sure all the brains know what hive they're in,
since they almost immediately end up holding a stale object.

If we need this, we should add it after some discussion