[!!!] For now, autocreate fb11admin user if no users
Please read the code here. As commented, this is here for now, until we
create a standard "set up your admin user" sort of UI, or pull the data
from the environment, or both.

Defaulting to fb11admin/fb11password is common, but we don't want to
accidentally leave a deployed system in this state. We'd rather at least
be able to set up the password up front.