Raise an exception if the parameter is not found in storage
This should make it easier to find bugs related to use of wrong names.

Note that I've had to special case in two places on whether 'namespace'
attribute is set, which I've done properly using strategy pattern and null
object.  However it's doubtful whether we need to support not having a namespace
at all, which means we should probably get rid of all that and make namespace
required instead.