Vendor in and port to fixed DBIC-DH VersionStorage::WithSchema
This tarball I'm vendoring in was built from the following pull
request's code:

https://github.com/Altreus/DBIx-Class-DeploymentHandler-VersionStorage-WithSchema/pull/5

When that's merged upstream and released to CPAN we can remove the
vendored version and rely on the version number specified in cpanfile.
For now we need to keep a tarball of the right version and install from
that.